Assessment/Module 16) – Python DB and Python Framework (Django Full Stack Development)/WriteSphere/requirements.txt
Django==4.2.10
PyMySQL==1.1.0
Pillow>=10.3.0
python-dotenv==1.0.0
django-ckeditor==6.7.0
django-taggit==5.0.1
whitenoise==6.6.0
django-allauth<65